requirements.txt
# ── Web 框架 ─────────────────────────────────
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
pydantic>=2.5,<3.0

# ── 資料處理 ─────────────────────────────────
pandas>=2.2,<4.0
numpy>=1.26,<3.0
yfinance>=0.2.40,<2.0

# ── LLM 客戶端 ───────────────────────────────
anthropic>=0.40,<2.0
google-genai>=1.0,<2.0
openai>=1.40,<3.0

# ── 工具 ─────────────────────────────────────
python-dotenv>=1.0,<2.0

# ── 開發/測試（可選；正式部署可拆 requirements-dev.txt） ──
pytest>=8.0,<10.0
